requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0
django-cors-headers>=4.3
channels>=4.0
channels-redis>=4.2
celery>=5.3
redis>=5.0
firebase-admin>=6.4
Pillow>=10.0
python-decouple>=3.8
daphne>=4.0
gunicorn>=22.0

dj-database-url>=2.1
psycopg2-binary>=2.9
